+++ CHANGELOG.md
@@ -114,4 +114,9 @@ All notable changes to this service will be documented in this file.
 
 ## [1.3.11]
 ### Notes
+- Added site-wide notification banner and feature flag
+- Set Entra SSO enabled to false
+
+## [1.3.12]
+### Notes
 - Make email address configurable on invite a contributor page